refactor(database_endpoints): moved db endpoints from main.py into separate files in database_endpoints folder
